Correct type issue in chip ID comparison logic

+++ src/XLR8Info.cpp
@@ -157,7 +157,7 @@ bool     XLR8Info::hasICSPVccGndSwap(void)  {
      0x38BF};
   uint32_t chipId = getChipId();
   for (int i=0;i< sizeof(cidTable)/sizeof(cidTable[0]);i++) {
-    uint32_t cidtoTest = (cidTable[i] << 16) + 0x6E00;
+    uint32_t cidtoTest = ((uint32_t)cidTable[i] << 16) + 0x6E00;
     if (chipId == cidtoTest) {return true;}
   }
   return false;
